backend/requirements.txt
bcrypt==4.0.1
email-validator==1.3.1
fastapi==0.95.0
httpx==0.23.3
itsdangerous==2.1.2
Jinja2==3.1.2
orjson==3.8.9
passlib==1.7.4
pydantic==1.10.7
PyJWT==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
SQLAlchemy==2.0.7
starlette==0.26.1
ujson==5.7.0
uvicorn==0.21.1
